chore(organize): cleanup to ng-sample directory to scale better
This brings some organization to the sample that should allow for
clearer growth paths to support more examples and test components.
