+++ ng-sample/LICENSE
