+++ ng-sample/app/main-page.ts
@@ -1,13 +1,13 @@
-import * as profiling from "./profiling";
+import * as profiling from "./performance/profiling";
 import {topmost} from 'ui/frame';
 import {TextView} from 'ui/text-view';
 import {Page} from 'ui/page';
 
 import {nativeScriptBootstrap} from './nativescript-angular/application';
-// import {RendererTest} from './renderer-test';
-//import {Benchmark} from './benchmark';
-//import {ListTest} from './list-test';
-import {ListTestAsync} from './list-test-async';
+// import {Benchmark} from './performance/benchmark';
+// import {RendererTest} from './examples/renderer-test';
+// import {ListTest} from './examples/list/list-test';
+import {ListTestAsync} from './examples/list/list-test-async';
 
 export function createPage() {
     var page = new Page();
@@ -18,9 +18,9 @@ export function createPage() {
 
         profiling.start('ng-bootstrap');
         console.log('BOOTSTRAPPING...');
-        //nativeScriptBootstrap(Benchmark, []).then((appRef) => {
+        // nativeScriptBootstrap(Benchmark, []).then((appRef) => {
         // nativeScriptBootstrap(RendererTest, []).then((appRef) => {
-        //nativeScriptBootstrap(ListTest, []).then((appRef) => {
+        // nativeScriptBootstrap(ListTest, []).then((appRef) => {
         nativeScriptBootstrap(ListTestAsync, []).then((appRef) => {
             profiling.stop('ng-bootstrap');
             console.log('ANGULAR BOOTSTRAP DONE.');
