+++ ng-sample/app/performance/benchmark.ts
