+++ ng-sample/app/performance/profiling.ts
